Move loss group and loss utils under engine subpackage for better package organization
